Training

+++ conv2d/model.py
@@ -19,50 +19,50 @@ def create_conv2d_model(data_in):
         (10 possible classes)
     """
     c1_1 = tf.layers.conv2d(inputs=data_in, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_1_1")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_1_1')
     c1_2 = tf.layers.conv2d(inputs=c1_1, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_1_2")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_1_2')
     c1_3 = tf.layers.conv2d(inputs=c1_2, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_1_3")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_1_3')
     c1_4 = tf.layers.conv2d(inputs=c1_3, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_1_4")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_1_4')
     p1 = tf.layers.max_pooling2d(inputs=c1_4, pool_size=2, strides=2,
-                                 padding="same", name="p2d_1")
+                                 padding='same', name='p2d_1')
 
     c2_1 = tf.layers.conv2d(inputs=p1, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_2_1")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_2_1')
     c2_2 = tf.layers.conv2d(inputs=c2_1, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_2_2")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_2_2')
     c2_3 = tf.layers.conv2d(inputs=c2_2, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_2_3")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_2_3')
     p2 = tf.layers.max_pooling2d(inputs=c2_3, pool_size=2, strides=2,
-                                 padding="same", name="p2d_2")
+                                 padding='same', name='p2d_2')
 
     c3_1 = tf.layers.conv2d(inputs=p2, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_3_1")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_3_1')
     c3_2 = tf.layers.conv2d(inputs=c3_1, filters=N_FILTERS, kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_3_2")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_3_2')
     p3 = tf.layers.max_pooling2d(inputs=c3_2, pool_size=2, strides=2,
-                                 padding="same", name="p2d_3")
+                                 padding='same', name='p2d_3')
 
     c4_1 = tf.layers.conv2d(inputs=p3, filters=(N_FILTERS / 2), kernel_size=4,
-                            padding="same", activation=tf.nn.relu,
-                            name="c2d_4_1")
+                            padding='same', activation=tf.nn.relu,
+                            name='c2d_4_1')
     p4 = tf.layers.max_pooling2d(inputs=c4_1, pool_size=2, strides=2,
-                                 padding="same", name="p2d_4")
+                                 padding='same', name='p2d_4')
 
     # p4.shape == (-1, 2, 2, N_FILTERS/2)
     # 2 == 32 / (2^4) where 32 = image width, height and 4 = pool layer count
-    p4_flattened = tf.reshape(p4, shape=[-1, 2 * 2 * (N_FILTERS / 2)])
+    p4_flattened = tf.reshape(p4, shape=[-1, 2 * 2 * int(N_FILTERS / 2)])
     d1 = tf.layers.dense(p4_flattened, units=2048, activation=tf.nn.relu,
                          use_bias=True)
     d2 = tf.layers.dense(d1, units=1024, activation=tf.nn.relu, use_bias=True)
